refactor: :recycle: refractor code and clean up functions
