+++ lib/remaths.dart
@@ -3,13 +3,16 @@ library remaths;
 import 'package:flutter/material.dart';
 import 'package:flutter/physics.dart';
 import 'dart:math' as math;
-import 'package:async/async.dart';
 
-export 'core/interpolateColors.dart';
-export 'core/measure.dart';
-
-part "core/animations/widget.dart";
-part "core/interpolate.dart";
-part "core/animations/spring.dart";
+part 'core/_intervals/_interpolate.dart';
 part 'core/node.dart';
 part 'core/operators.dart';
+part 'core/_intervals/ops.dart';
+
+part 'core/_intervals/_funcs.dart';
+part 'core/_intervals/_operators.dart';
+part 'core/_intervals/_shared.dart';
+part 'core/_intervals/_spring.dart';
+part 'core/_intervals/_interpolateColors.dart';
+part 'core/measure.dart';
+part 'core/_intervals/_measure.dart';
